td5

+++ MAPC/TD5/src/main/java/org/example/Beverage.java
@@ -2,5 +2,5 @@ package org.example;
 
 public interface Beverage {
     public String getDescription();
-    public Integer cost();
+    public float cost();
 }
